+++ AStar.h
@@ -31,6 +31,7 @@
 #define AStar_h
 
 #include <stdlib.h>
+#include <stdint.h>
 
 typedef struct __ASNeighborList *ASNeighborList;
 typedef struct __ASPath *ASPath;
@@ -38,13 +39,13 @@ typedef struct __ASPath *ASPath;
 typedef struct {
     size_t  nodeSize;                                                                               // the size of the structure being used for the nodes - important since nodes are copied into the resulting path
     void    (*nodeNeighbors)(ASNeighborList neighbors, void *node, void *from_node, void *context);                  // add nodes to the neighbor list if they are connected to this node
-    float   (*pathCostHeuristic)(void *fromNode, void *toNode, void *context);                      // estimated cost to transition from the first node to the second node -- optional, uses 0 if not specified
+    uint64_t   (*pathCostHeuristic)(void *fromNode, void *toNode, void *context);                      // estimated cost to transition from the first node to the second node -- optional, uses 0 if not specified
     int     (*earlyExit)(size_t visitedCount, void *visitingNode, void *goalNode, void *context);   // early termination, return 1 for success, -1 for failure, 0 to continue searching -- optional
     int     (*nodeComparator)(void *node1, void *node2, void *context);                             // must return a sort order for the nodes (-1, 0, 1) -- optional, uses memcmp if not specified
 } ASPathNodeSource;
 
 // use in the nodeNeighbors callback to return neighbors
-void ASNeighborListAdd(ASNeighborList neighbors, void *node, float edgeCost);
+void ASNeighborListAdd(ASNeighborList neighbors, void *node, uint64_t edgeCost);
 
 // if goalNode is NULL, it searches the entire graph and returns the cheapest deepest path
 // context is optional and is simply passed through to the callback functions
@@ -60,7 +61,7 @@ void ASPathDestroy(ASPath path);
 ASPath ASPathCopy(ASPath path);
 
 // fetches the total cost of the path
-float ASPathGetCost(ASPath path, size_t i);
+uint64_t ASPathGetCost(ASPath path, size_t i);
 
 // fetches the number of nodes in the path
 size_t ASPathGetCount(ASPath path);
